% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{vocabulary_counts}
\alias{vocabulary_counts}
\title{vocabulary counts}
\usage{
vocabulary_counts(train_data = NULL, MAX_vocab = 0, MIN_count = 1,
  output_vocabulary = NULL, trace = FALSE)
}
\arguments{
\item{train_data}{a character string specifying the path to the train text file}

\item{MAX_vocab}{a value specifying the number of terms in the vocabulary. For instance a \emph{MAX_vocab} value of 0 includes all the vocab-terms.}

\item{MIN_count}{a value greater or equal to 1. It specifies the minimum occurrences (counts of words) for inclusion in the vocabulary}

\item{output_vocabulary}{a character string specifying the path to the output text file}

\item{trace}{either TRUE or FALSE. If TRUE information will be printed out}
}
\value{
a character string specifying the location of the saved data
}
\description{
vocabulary counts
}
\examples{

# library(GloveR)

# res = vocabulary_counts(train_data = '/data_GloveR/dat.txt', MAX_vocab = 0,

#                         MIN_count = 5, output_vocabulary = '/data_GloveR/VOCAB.txt', trace = T)
}
\references{
https://github.com/stanfordnlp/GloVe

http://nlp.stanford.edu/projects/glove/

http://nlp.stanford.edu/pubs/glove.pdf
}

